api/requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/dev.txt
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

appdirs==1.4.4
    # via
    #   -r requirements/dev.txt
    #   black
asgiref==3.4.1
    # via
    #   -r requirements/dev.txt
    #   django
attrs==21.2.0
    # via
    #   -r requirements/dev.txt
    #   pytest
backcall==0.2.0
    # via
    #   -r requirements/dev.txt
    #   ipython
black==21.6b0
    # via -r requirements/dev.txt
certifi==2021.5.30
    # via
    #   -r requirements/dev.txt
    #   requests
chardet==4.0.0
    # via
    #   -r requirements/dev.txt
    #   requests
click==8.0.1
    # via
    #   -r requirements/dev.txt
    #   black
    #   pip-tools
coreapi==2.3.3
    # via
    #   -r requirements/dev.txt
    #   drf-yasg
coreschema==0.0.4
    # via
    #   -r requirements/dev.txt
    #   coreapi
    #   drf-yasg
decorator==5.0.9
    # via
    #   -r requirements/dev.txt
    #   ipython
django==3.2.5
    # via
    #   -r requirements/dev.txt
    #   django-extensions
    #   django-filter
    #   djangorestframework
    #   drf-yasg
django-environ==0.4.5
    # via -r requirements/dev.txt
django-extensions==3.1.3
    # via -r requirements/dev.txt
django-filter==2.4.0
    # via -r requirements/dev.txt
djangorestframework==3.12.4
    # via
    #   -r requirements/dev.txt
    #   djangorestframework-gis
    #   drf-yasg
djangorestframework-gis==0.17
    # via -r requirements/dev.txt
drf-yasg==1.20.0
    # via -r requirements/dev.txt
gunicorn==20.1.0
    # via -r requirements/dev.txt
idna==2.10
    # via
    #   -r requirements/dev.txt
    #   requests
inflection==0.5.1
    # via
    #   -r requirements/dev.txt
    #   drf-yasg
iniconfig==1.1.1
    # via
    #   -r requirements/dev.txt
    #   pytest
ipython==7.25.0
    # via -r requirements/dev.txt
ipython-genutils==0.2.0
    # via
    #   -r requirements/dev.txt
    #   traitlets
itypes==1.2.0
    # via
    #   -r requirements/dev.txt
    #   coreapi
jedi==0.18.0
    # via
    #   -r requirements/dev.txt
    #   ipython
jinja2==3.0.1
    # via
    #   -r requirements/dev.txt
    #   coreschema
markupsafe==2.0.1
    # via
    #   -r requirements/dev.txt
    #   jinja2
matplotlib-inline==0.1.2
    # via
    #   -r requirements/dev.txt
    #   ipython
mypy-extensions==0.4.3
    # via
    #   -r requirements/dev.txt
    #   black
packaging==21.0
    # via
    #   -r requirements/dev.txt
    #   drf-yasg
    #   pytest
parso==0.8.2
    # via
    #   -r requirements/dev.txt
    #   jedi
pathspec==0.8.1
    # via
    #   -r requirements/dev.txt
    #   black
pep517==0.10.0
    # via
    #   -r requirements/dev.txt
    #   pip-tools
pexpect==4.8.0
    # via
    #   -r requirements/dev.txt
    #   ipython
pickleshare==0.7.5
    # via
    #   -r requirements/dev.txt
    #   ipython
pip-tools==6.2.0
    # via -r requirements/dev.txt
pluggy==0.13.1
    # via
    #   -r requirements/dev.txt
    #   pytest
prompt-toolkit==3.0.19
    # via
    #   -r requirements/dev.txt
    #   ipython
psycopg2-binary==2.9.1
    # via -r requirements/dev.txt
ptyprocess==0.7.0
    # via
    #   -r requirements/dev.txt
    #   pexpect
py==1.10.0
    # via
    #   -r requirements/dev.txt
    #   pytest
pygments==2.9.0
    # via
    #   -r requirements/dev.txt
    #   ipython
pyparsing==2.4.7
    # via
    #   -r requirements/dev.txt
    #   packaging
pytest==6.2.4
    # via
    #   -r requirements/dev.txt
    #   pytest-django
pytest-django==4.4.0
    # via -r requirements/dev.txt
pytz==2021.1
    # via
    #   -r requirements/dev.txt
    #   django
regex==2021.7.6
    # via
    #   -r requirements/dev.txt
    #   black
requests==2.25.1
    # via
    #   -r requirements/dev.txt
    #   coreapi
ruamel.yaml==0.17.10
    # via
    #   -r requirements/dev.txt
    #   drf-yasg
ruamel.yaml.clib==0.2.6
    # via
    #   -r requirements/dev.txt
    #   ruamel.yaml
sqlparse==0.4.1
    # via
    #   -r requirements/dev.txt
    #   django
toml==0.10.2
    # via
    #   -r requirements/dev.txt
    #   black
    #   pep517
    #   pytest
traitlets==5.0.5
    # via
    #   -r requirements/dev.txt
    #   ipython
    #   matplotlib-inline
uritemplate==3.0.1
    # via
    #   -r requirements/dev.txt
    #   coreapi
    #   drf-yasg
urllib3==1.26.6
    # via
    #   -r requirements/dev.txt
    #   requests
wcwidth==0.2.5
    # via
    #   -r requirements/dev.txt
    #   prompt-toolkit
wheel==0.36.2
    # via
    #   -r requirements/dev.txt
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
